Relocate existing tests for appropriate discoverability.
1. Add a __init__.py to make napari/_tests packaged.
2. Rename existing `tests` directories `_tests` to keep them out of the public API.

Fixes #875
