+++ napari/_qt/_tests/test_qt_play.py
