+++ napari/_qt/_tests/test_qt_range_slider_popup.py
