+++ napari/_qt/layers/_tests/test_qt_image_base_layer_.py
