+++ napari/_tests/test_advanced.py
