+++ napari/_tests/test_keybindings.py
