+++ napari/_tests/test_mouse_bindings.py
