+++ napari/_tests/test_numpy_like.py
