+++ napari/_tests/test_viewer_layer_parity.py
