+++ napari/_tests/utils.py
