+++ napari/_vispy/_tests/test_image_rendering.py
