+++ napari/_vispy/_tests/test_vispy_big_images.py
