+++ napari/components/_tests/test_dims.py
