+++ napari/components/_tests/test_layers_list.py
