+++ napari/components/_tests/test_multichannel.py
