+++ napari/components/_tests/test_viewer_image_io.py
