+++ napari/layers/_tests/test_layer_utils.py
