+++ napari/layers/image/_tests/test_big_image_timing.py
