+++ napari/layers/image/_tests/test_image.py
