+++ napari/layers/image/_tests/test_image_utils.py
