+++ napari/layers/image/_tests/test_pyramid.py
