+++ napari/layers/image/_tests/test_volume.py
