+++ napari/layers/labels/_tests/test_labels.py
