+++ napari/layers/labels/_tests/test_pyramid.py
