+++ napari/layers/points/_tests/test_points.py
