+++ napari/layers/shapes/_tests/test_shape_list.py
