+++ napari/layers/shapes/_tests/test_shapes.py
