+++ napari/layers/shapes/_tests/test_shapes_keybindings.py
