+++ napari/layers/shapes/shape_models/_tests/test_shape_models.py
