+++ napari/layers/vectors/_tests/test_vectors.py
