+++ napari/utils/_tests/colors_data.py
