+++ napari/utils/_tests/test_color_to_array.py
