+++ napari/utils/_tests/test_colormaps.py
