+++ napari/utils/_tests/test_info.py
