+++ napari/utils/_tests/test_interactions.py
