+++ napari/utils/_tests/test_keybindings.py
