+++ napari/utils/_tests/test_misc.py
