+++ napari/utils/_tests/test_naming.py
