+++ napari/utils/_tests/test_validators.py
